refactor: add base package
