+++ infrastructure/src/main/java/com/cleanarch/infrastructure/CleanArchitecture.java
@@ -1,4 +1,4 @@
-package com.cleanarch;
+package com.cleanarch.infrastructure;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
