+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/entity/UserEntity.java
@@ -1,4 +1,4 @@
-package com.cleanarch.persistence.entity;
+package com.cleanarch.infrastructure.persistence.entity;
 
 import com.cleanarch.core.domain.enums.UserTypeEnum;
 import jakarta.persistence.*;
