+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/entity/WalletEntity.java
@@ -1,4 +1,4 @@
-package com.cleanarch.persistence.entity;
+package com.cleanarch.infrastructure.persistence.entity;
 
 import jakarta.persistence.*;
