+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/repository/TransactionPinRepository.java
@@ -1,6 +1,6 @@
-package com.cleanarch.persistence.repository;
+package com.cleanarch.infrastructure.persistence.repository;
 
-import com.cleanarch.persistence.entity.TransactionPinEntity;
+import com.cleanarch.infrastructure.persistence.entity.TransactionPinEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
