+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/repository/TransactionRepository.java
@@ -1,4 +1,4 @@
-package com.cleanarch.persistence.repository;
+package com.cleanarch.infrastructure.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
