+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.cleanarch.persistence.repository;
+package com.cleanarch.infrastructure.persistence.repository;
 
-import com.cleanarch.persistence.entity.UserEntity;
+import com.cleanarch.infrastructure.persistence.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
