+++ infrastructure/src/main/java/com/cleanarch/infrastructure/persistence/repository/WalletRepository.java
@@ -1,6 +1,6 @@
-package com.cleanarch.persistence.repository;
+package com.cleanarch.infrastructure.persistence.repository;
 
-import com.cleanarch.persistence.entity.WalletEntity;
+import com.cleanarch.infrastructure.persistence.entity.WalletEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
